// src/views/Lists/Chemicals/interfaces.ts
export interface ModalTypes {
  RECETEASAMAID?: string;
  BOYATIPIID?: string;
  BOYATIPIADI?: string;
  BANYOCARPANNO?: string;
  DOZAJLAMAPROGRAMNO?: string;
  SURE?: string;
  ORGATEXETEKBASINAAKTARILIR?: string;
  TUZSULFATBASTAN?: string;
  ADI?: string;
  BOYAASAMASIMI?: boolean;
}

export interface IModalForm {
  ADI?: string;
  BOYATIPIID?: string;
  BOYATIPIADI?: string;
  BANYOCARPANNO?: string;
  DOZAJLAMAPROGRAMNO?: string;
  SURE?: string;
  ORGATEXETEKBASINAAKTARILIR?: string;
  TUZSULFATBASTAN?: string;
  BOYAASAMASIMI?: boolean;
  RECETEASAMAID?: string;
}